requirements.txt
Django>=4.2,<5.0
google-generativeai>=0.5.0
python-dotenv>=1.0.0
Pillow>=10.0.0
whitenoise>=6.5.0
gunicorn>=21.2.0
